requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
pandas>=2.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
requests>=2.31.0
openai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
pymongo>=4.6.0
opencv-python>=4.8.0
pillow>=10.0.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
